Mark read and unread

+++ app/inbox/views.py
@@ -15,6 +15,7 @@ def inbox(request):
         for i in all_notifs:
             new_notif = i.to_standard_dict()
             new_notif['created_on'] = i.created_on
+            new_notif['modified_on'] = i.modified_on
             new_notif['username'] = get_user_model().objects.get(id=new_notif['from_user_id']).username
 
             params.append(new_notif)
@@ -31,13 +32,61 @@ def delete_notification(request):
     if request.method == 'DELETE' and profile is not None:
         params = dict()
         params['success'] = []
-        for i in json.loads(request.body.decode('utf-8'))['delete']:
-            entry = Notification.objects.filter(id=i)
-            if entry.to_user_id == request.user.id and len(entry) != 0:
-                entry.delete()
-                params['success'].append(True)
-            else:
-                params['success'].append(False)
+        req_body = json.loads(request.body.decode('utf-8'))
+        if 'delete' in req_body:
+            for i in req_body['delete']:
+                entry = Notification.objects.filter(id=i)
+                if entry.to_user_id == request.user.id and len(entry) != 0:
+                    entry.delete()
+                    params['success'].append(True)
+                else:
+                    params['success'].append(False)
+        return JsonResponse(params, status=200)
+
+    else:
+        return HttpResponseForbidden('Not Allowed')
+
+
+def unread_notification(request):
+    """Mark a notification as unread"""
+    profile = request.user.profile if request.user.is_authenticated and request.user.profile else None
+    if request.method == 'PUT' and profile is not None:
+        params = dict()
+        params['success'] = []
+        req_body = json.loads(request.body.decode('utf-8'))
+        if 'unread' in req_body:
+            for i in req_body['unread']:
+                entry = Notification.objects.filter(id=i)
+                if entry.to_user_id == request.user.id and len(entry) != 0:
+                    obj = entry[0]
+                    obj.is_read = False
+                    obj.save()
+                    params['success'].append(True)
+                else:
+                    params['success'].append(False)
+        return JsonResponse(params, status=200)
+
+    else:
+        return HttpResponseForbidden('Not Allowed')
+
+
+def read_notification(request):
+    """Mark a notification as unread"""
+    profile = request.user.profile if request.user.is_authenticated and request.user.profile else None
+    if request.method == 'PUT' and profile is not None:
+        params = dict()
+        params['success'] = []
+        req_body = json.loads(request.body.decode('utf-8'))
+        if 'read' in req_body:
+            for i in req_body['unread']:
+                entry = Notification.objects.filter(id=i)
+                if entry.to_user_id == request.user.id and len(entry) != 0:
+                    obj = entry[0]
+                    obj.is_read = True
+                    obj.save()
+                    params['success'].append(True)
+                else:
+                    params['success'].append(False)
         return JsonResponse(params, status=200)
 
     else:
